requirements.txt
semantic-kernel
python-dotenv
aiohttp
streamlit
openai
numpy
pandas